pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.2.0",
    "modal>=1.3.1",
    "typer>=0.21.1",
    "pyyaml>=6.0.3",
    "openai>=2.16.0",
    "anthropic>=0.76.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.0",
    "jupyterlab>=4.4.10",
    "notebook>=7.4.7",
    "pytest>=8.4.2",
]
